remove arbitrary styles from Nav page
Problem: the Nav page contained arbitrary styles for native elements, which is against our new rules for look and feel. These rules are moved to the declarative Shadow DOM of the default composition.

I could not go around a problem of too little specifity of the `ul` selector so I added !important. Better ideas are welcome.